+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ## History
 
-### 0.1.0 @2022 Aug xx
+### 0.1.1 @2022 Aug xx
+
+ - rename: `FileInfo` to `File`
+ - new   : `File` => `getDir` > returns directory, file's return parent directory
+ - update: `constructor` if no argument supplied the current directory is used
+ - update: `getFile` & `getFiles` return `File` object(s)
+ - minor : Improvements & fixes
+
+### 0.1.0 @2022 Aug 04
 
  - initial